requirements.txt
fastapi==0.112.2  
uvicorn[standard]==0.30.6  
python-dotenv==1.0.1  
Pillow==10.4.0  
pydantic==2.8.2  
google-genai==0.3.0  
